requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
aiohttp==3.11.0
# transmission-rpc==7.0.11
python-multipart==0.0.6
aiogram==3.17.0
# asyncpg==0.30.0
python-dotenv==1.0.0
# sqlalchemy==2.0.37
# bencodepy==0.9.5